requirements.txt
langcodes
language_data

tqdm
python-dotenv

requests
google-cloud-translate

torch
stanza
transformers
nemo_toolkit[all]

sacrebleu
rouge
bert-score
unbabel-comet

hydra
omegaconf